backend/requirements/test-requirements.txt
check-jsonschema
httpx
pytest
pytest-ordering
pytest-asyncio
pytest-mock
pyright
ruff
mypy
types-aiofiles
types-ldap3
types-jsonschema
types-psutil
types-requests

